+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export {EzTimeline} from './timeline/ez-timeline';
 export {EzGenealogy} from './genealogy/ez-genealogy';
+export {EzMap} from './map/ez-map';